requirements.txt
# Python dependencies untuk testing notification system

# HTTP requests
requests>=2.31.0

# Environment variables
python-dotenv>=1.0.0
